Calling AllAuctions for console app as well
