+++ surplus-auctioneer-webapp/WebHelpers/Tools.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+using surplus_auctioneer_models;
+using surplus_auctioneer_webdata;
+
+namespace surplus_auctioneer_webapp.WebHelpers
+{
+    public static class Tools
+    {
+        
+    }
+}
