+++ surplus-auctioneer-webapp/surplus-auctioneer-webapp.csproj
@@ -207,8 +207,8 @@
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
-    <Compile Include="Helpers\IsNumericAttribute.cs" />
-    <Compile Include="Helpers\Tools.cs" />
+    <Compile Include="WebHelpers\IsNumericAttribute.cs" />
+    <Compile Include="WebHelpers\Tools.cs" />
     <Compile Include="Models\AccountViewModels.cs" />
     <Compile Include="Models\AuctionViewModel.cs" />
     <Compile Include="Models\IdentityModels.cs" />
