+++ surplus-auctioneer-webdata/WisconsinAuctionData.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using System.Net;
 using System.Text.RegularExpressions;
 using surplus_auctioneer_models;
-using wi_auctioneer_webdata;
 using HAP = HtmlAgilityPack;
 
 namespace surplus_auctioneer_webdata
